lifting the state up, and passing down reference.
Products Control, needs to pass to ProductManager that it is pressed, ProductManager updates this dynamically (it is stateful) and then add it to the Products.

To call addProduct, whenver the button is pressed, but the button lives in product_control. So to pass a reference down to this object..

BIG TAKEWAY:  To pass data up to ParentControl, from a button living in a differnt widget. To pass data UP, and the parent widget needs about it, we do it by passing a reference DOWN